fix: refactor renderers
